processing working. Not tested in docker yet
