+++ app/models/__init__.py
@@ -1,2 +1,1 @@
 from .models import *
-from .app_user import *
