+++ app/views/tasks/view.py
@@ -7,6 +7,7 @@ from app.models import Task, TaskSchema, db, TaskStatus
 from flask_restful import Resource
 from werkzeug.utils import secure_filename
 from flask_jwt_extended import jwt_required
+from ...tareas import process_video
 
 video_schema = TaskSchema()
 
@@ -22,6 +23,7 @@ def getTasks():
 
 
 @task_blueprint.route('', methods=['POST'])
+@jwt_required()
 def createTask():
     if 'file' not in request.files:
         return {'message': 'No file part'}, 400
@@ -39,6 +41,7 @@ def createTask():
             filename=filename, timestamp=datetime.now(), status=TaskStatus.UPLOADED)
         db.session.add(new_video)
         db.session.commit()
+        process_video.delay(filename)
         return video_schema.dump(new_video), 201
 
 
@@ -47,6 +50,9 @@ def getTask(id_task):
     task = Task.query.get(id_task)
     if not task:
         return {'message': 'Task not found'}, 404
+    
+    if task.status != TaskStatus.PROCESSED:
+        return {'message': 'Task not processed yet'}, 400
 
     processed_file_url = f'/api/tasks/{id_task}/processed'
