+++ docker-compose.yml
@@ -13,6 +13,8 @@ services:
       FLASK_ENV: development
     depends_on:
       - db
+      - redis
+      - celery
   db:
     image: postgres:latest
     environment:
@@ -29,6 +31,17 @@ services:
       - "5050:80"
     depends_on:
       - db
+
+  redis:
+    image: redis:latest
+
+  celery:
+    image: python:3.9-slim
+    command: /bin/sh -c "pip install -r /app/requirements.txt" && celery -A app.tareas.tareas worker -l info
+    volumes:
+      - .:/app
+    depends_on:
+      - redis
 
 volumes:
   postgres-data:
