+++ requirements.txt
@@ -7,5 +7,6 @@ marshmallow
 marshmallow_sqlalchemy
 Werkzeug
 SQLAlchemy
-
-
+celery
+redis
+moviepy
